Add failure_recovery_tests.rs
